feat:refactored part of code and modify predicates generation

+++ src/main/java/ddfinder/evidence/IClueOffset.java
@@ -10,6 +10,6 @@ import java.util.List;
 public interface IClueOffset {
     final double ERR = 0.000000001;
     int[] countDouble(IPli probePli, int isSingle, Double[] keys, int startPos, double key, List<Double> thresholds);
-    int[] countInt(IPli probePli, int isSingle, Integer[] keys, int startPos, int key, List<Double> thresholds);
+    int[] countInt(IPli probePli, int isSingle, Long[] keys, int startPos, long key, List<Double> thresholds);
 
 }
